requirements.txt
fastapi>=0.103.0
sqlalchemy~=2.0.22
asyncio>=3.4.3
python-dotenv>=1.0.0
email-mime-text>=1.0.0
smtplib>=3.6.2
uvicorn~=0.24.0
asyncpg~=0.30.0
python-dotenv~=1.0.1
python-jose[cryptography]~=3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]~=2.11.7
